1.9

+++ src/main/java/net/glowstone/net/codec/play/entity/EntityTeleportCodec.java
@@ -11,9 +11,9 @@ public final class EntityTeleportCodec implements Codec<EntityTeleportMessage> {
     @Override
     public EntityTeleportMessage decode(ByteBuf buf) throws IOException {
         int id = ByteBufUtils.readVarInt(buf);
-        int x = buf.readInt();
-        int y = buf.readInt();
-        int z = buf.readInt();
+        double x = buf.readDouble();
+        double y = buf.readDouble();
+        double z = buf.readDouble();
         int rotation = buf.readByte();
         int pitch = buf.readByte();
         boolean ground = buf.readBoolean();
@@ -23,9 +23,9 @@ public final class EntityTeleportCodec implements Codec<EntityTeleportMessage> {
     @Override
     public ByteBuf encode(ByteBuf buf, EntityTeleportMessage message) throws IOException {
         ByteBufUtils.writeVarInt(buf, message.getId());
-        buf.writeInt(message.getX());
-        buf.writeInt(message.getY());
-        buf.writeInt(message.getZ());
+        buf.writeDouble(message.getX());
+        buf.writeDouble(message.getY());
+        buf.writeDouble(message.getZ());
         buf.writeByte(message.getRotation());
         buf.writeByte(message.getPitch());
         buf.writeBoolean(message.isOnGround());
